Web/Django/tutorial/requirements.txt
Django
djangorestframework
django_filter